Constructed Question class, model part of MVC
Constructed our Question class, which is the model part
of our MVC Architecture. It is a blue print of a question
displayed in our MainActivity (our view). Contains attributes
for question answer id and answer being true or false. This
data for the model is retrieved inside strings.xml resource.
 Created getters and setters for the two data members too