Added next question button and counter
Created an ImageButton nextButton, connected
to nextButton's Id in the view, set on click
listener for nextButton. So, when nextButton
is clicked, increment the counter by 1, then
perform modulo to get the remainder to ensure
when the current question counter equals the
last question in the questionBank, the counter
loops back to 0. At each current question index,
set the current question text to the current
question we are at in the question bank by index
and then grab the answerResId data member.