Added app functionality to check if users answer is correct
First refactored code in switch case for next button,
so when user clicks on next button, the code calls on
updateQuestion(). Refactored code in false_button and
true_button case, so code calls on checkAnswer()
method to check is the answer the user chose the same
as the actual correct answer. Based on result of this
check, app will display a toast message indicating
whether user's chosen answer was actually correct by
displaying That's correct or That's incorrect.